refactor(browser): extract lightweight protocol crate
