+++ crates/horizon-browser-protocol/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "horizon-browser-protocol"
+publish = false
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+description = "Lightweight wire types for Horizon browser automation"
+repository.workspace = true
+readme = "README.md"
+
+[lints]
+workspace = true
+
+[dependencies]
+serde.workspace = true
+serde_json.workspace = true
