+++ crates/horizon-browser/src/lib.rs
@@ -7,7 +7,6 @@
 
 mod audit;
 pub mod cdp;
-mod control;
 mod coordination;
 mod disclosure;
 mod error;
@@ -25,14 +24,14 @@ mod websocket;
 use std::path::PathBuf;
 
 pub use audit::{BrowserAuditAction, BrowserAuditActor, BrowserAuditEntry, BrowserAuditStatus, new_action_id};
-pub use control::{
-    AgentAction, BrowserControlAction, DEFAULT_CLICK_COUNT, MAX_CLICK_COUNT, MAX_QUERY_RESULTS, MAX_SNAPSHOT_NODES,
-    normalize_navigation_target,
-};
 pub use coordination::{BrowserCoordination, CoordinationSignals, CoordinationState, HandoffRequest};
 pub use disclosure::{AutomationDisclosurePolicy, AutomationDisclosureStatus};
 pub use error::BrowserError;
 pub use frames::{FrameData, FrameMetrics, FrameSlot, PageScrollState};
+pub use horizon_browser_protocol::{
+    AgentAction, BackendAvailability, BackendCapabilities, BackendKind, BrowserControlAction, DEFAULT_CLICK_COUNT,
+    FrameDelivery, MAX_CLICK_COUNT, MAX_QUERY_RESULTS, MAX_SNAPSHOT_NODES, normalize_navigation_target,
+};
 pub use input::{BrowserButton, BrowserEditCommand, BrowserInput, BrowserKey, BrowserModifiers};
 pub use network::{
     BrowserNetworkCapture, BrowserNetworkCaptureOptions, BrowserNetworkConnection, BrowserNetworkConnectionState,
@@ -110,44 +109,6 @@ impl BrowserConfig {
     }
 }
 
-/// Browser automation backend selected for a session.
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq, serde::Deserialize, serde::Serialize)]
-pub enum BackendKind {
-    #[default]
-    #[serde(rename = "chromium", alias = "chromium-cdp")]
-    ChromiumCdp,
-    #[serde(rename = "firefox", alias = "firefox-bidi")]
-    FirefoxBidi,
-    #[serde(rename = "safari", alias = "safari-web-driver")]
-    SafariWebDriver,
-}
-
-/// How a backend can provide pixels to the caller.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum FrameDelivery {
-    PushJpeg,
-    AdaptiveScreenshot,
-    RecordingFile,
-    Unsupported,
-}
-
-/// Exact behavior callers may rely on for a selected backend.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-#[allow(clippy::struct_excessive_bools)] // independent protocol feature claims
-pub struct BackendCapabilities {
-    pub frame_delivery: FrameDelivery,
-    pub viewport: bool,
-    pub physical_keys: bool,
-    pub ime: bool,
-    pub clipboard: bool,
-    pub downloads: bool,
-    pub persistent_profile: bool,
-    /// Whether the backend can install common-signal minimization before
-    /// author scripts run.
-    pub automation_disclosure_minimization: bool,
-    pub max_sessions: Option<u32>,
-}
-
 /// Runtime extensions negotiated for one active backend session.
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub struct ActiveBackendCapabilities {
@@ -155,74 +116,6 @@ pub struct ActiveBackendCapabilities {
     pub capabilities: BackendCapabilities,
     pub bidi: bool,
     pub automation_disclosure: AutomationDisclosureStatus,
-}
-
-/// Compile-time host support for a backend. Executable discovery remains a
-/// launch-time check because installations can change while Horizon runs.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum BackendAvailability {
-    Available,
-    UnsupportedPlatform(&'static str),
-}
-
-impl BackendKind {
-    #[must_use]
-    pub const fn capabilities(self) -> BackendCapabilities {
-        match self {
-            Self::ChromiumCdp => BackendCapabilities {
-                frame_delivery: FrameDelivery::PushJpeg,
-                viewport: true,
-                physical_keys: true,
-                ime: true,
-                clipboard: true,
-                downloads: false,
-                persistent_profile: true,
-                automation_disclosure_minimization: true,
-                max_sessions: None,
-            },
-            Self::FirefoxBidi => BackendCapabilities {
-                frame_delivery: FrameDelivery::AdaptiveScreenshot,
-                viewport: true,
-                physical_keys: false,
-                ime: true,
-                clipboard: false,
-                downloads: false,
-                persistent_profile: true,
-                automation_disclosure_minimization: true,
-                max_sessions: None,
-            },
-            Self::SafariWebDriver => BackendCapabilities {
-                frame_delivery: FrameDelivery::AdaptiveScreenshot,
-                viewport: true,
-                physical_keys: false,
-                ime: false,
-                clipboard: false,
-                downloads: false,
-                persistent_profile: false,
-                automation_disclosure_minimization: false,
-                max_sessions: Some(1),
-            },
-        }
-    }
-
-    #[must_use]
-    pub const fn display_name(self) -> &'static str {
-        match self {
-            Self::ChromiumCdp => "Chromium",
-            Self::FirefoxBidi => "Firefox",
-            Self::SafariWebDriver => "Safari",
-        }
-    }
-
-    #[must_use]
-    pub const fn availability(self) -> BackendAvailability {
-        match self {
-            Self::SafariWebDriver if !cfg!(target_os = "macos") => {
-                BackendAvailability::UnsupportedPlatform("Safari automation is available only on macOS")
-            }
-            _ => BackendAvailability::Available,
-        }
-    }
 }
 
 #[cfg(test)]
